+++ sokannonser/rest/model/result_models.py
@@ -35,11 +35,6 @@ suggest_typeahead_item = ns_platsannons.model('TypeaheadItem', {
     'type': fields.String(),
 })
 
-suggest_typeahead_results = ns_platsannons.model('TypeaheadResults', {
-    'time_in_millis': fields.Integer(),
-    'typeahead': fields.List(fields.Nested(suggest_typeahead_item))
-})
-
 
 class AdUrl(fields.Raw):
     def format(self, value):