+++ tests/api_tests/search/test_platsannonser_complete.py
@@ -41,8 +41,6 @@ def test_complete_endpoint_synonyms_typeahead(session, search_url, query, synony
     ('angu', ['angular', 'angularjs', 'angular.js']),
     ('pyth', ['python', 'python stockholm', 'python stockholms län', 'python göteborg', 'python västra götaland',
               'python västra götalands län']),
-    ('#coro', ['creo', 'foto', 'core', 'code', 'cura']),
-    ('#coron', ['fordon', 'core']),
     ('c#', ['c#']),
     ('c+', ['c++']),
     ('ang', ['angular', 'angularjs', 'angular.js', 'angered']),
@@ -113,28 +111,123 @@ def test_complete_multiple_words(session, search_url, query, query_2, expected_t
     compare_typeahead(typeahead, expected_typeahead)
 
 
-@pytest.mark.parametrize("contextual,expected", [(True,
-                                                  ['malmö butikssäljare',
-                                                   'malmö sjuksköterska',
-                                                   'malmö civilingenjör',
-                                                   'malmö högskoleingenjör',
-                                                   'malmö lagerarbetare',
-                                                   'malmö redovisningsekonom']),
-                                                 (False, ['malmö butikssäljare',
-                                                          'malmö sjuksköterska',
-                                                          'malmö civilingenjör',
-                                                          'malmö högskoleingenjör',
-                                                          'malmö lagerarbetare',
-                                                          'malmö redovisningsekonom'])])
-def test_complete_for_locations_with_space_and_contextual_param(session, search_url, contextual, expected):
+@pytest.mark.parametrize("contextual", [True, False])
+def test_complete_for_locations_with_space_and_contextual_param(session, search_url, contextual):
     """
     Test typeahead for location with trailing space after city name,
     and using parameter 'contextual' True or False
     """
+
     query = 'Malmö '
+    expected = ['malmö butikssäljare', 'malmö sjuksköterska', 'malmö civilingenjör', 'malmö högskoleingenjör',
+                'malmö lagerarbetare', 'malmö redovisningsekonom']
     response_json = get_complete(session, search_url, {'q': query, 'limit': 10, 'contextual': contextual})
     typeahead = response_json['typeahead']
     suggestions = []
     for t in typeahead:
         suggestions.append(t['value'])
     assert suggestions == expected
+
+
+@pytest.mark.parametrize("query, expected", [
+    ("göteborg sjuksköterska", ['göteborg sjuksköterska']),
+    ("götteborg sjuksköterska", ['göteborg sjuksköterska']),
+    ("götteborg sjukssköterska", ['göteborg sjuksköterska', 'göteborg sjuksköterskan']),
+    ("göteborg sjukssköterska", ['göteborg sjuksköterska', 'göteborg sjuksköterskan']),
+    ("göteborg sjukssköterska läckare", ['göteborg sjuksköterska läkare', 'göteborg sjuksköterska lärare']),
+    ("göteborg sjukssköterska läkkare", ['göteborg sjuksköterska läkare', 'göteborg sjuksköterska lärare']),
+    ("göteborg sjukssköterska lääkare", ['göteborg sjuksköterska läkare', 'göteborg sjuksköterska lärare']),
+    ("göteborg sjukssköterska läkare", ['göteborg sjuksköterska läkare', 'göteborg sjuksköterska lärare']),
+    ("läckare", ['läkare']),
+    ("götteborg", ['göteborg']),
+    ("stokholm", ['stockholm', 'stockholm city']),
+    ("stokholm ", ['stockholm', 'stockholm city']),  # trailing space
+    ("stockhlm", ['stockholm', 'stockholm city']),
+    ("stockhlm ", ['stockholm city']),  # trailing space
+    ("stokholm lärarre", ['stockholms lärare', 'stockholm lärare', 'stockholm läkare']),
+    ("göteborg sjukssköterska läckare", ['göteborg sjuksköterska lärare', 'göteborg sjuksköterska läkare']),
+    ("göteborg läckare sjukssköterska ", ['göteborg lärare sjuksköterska', 'göteborg läkare sjuksköterska']),
+    ("göteborg läckare sjukssköterska", ['göteborg lärare sjuksköterska', 'göteborg läkare sjuksköterska']),
+    ("läckare götteborg", ['läkare göteborg']),
+    ("läckare göteborg", ['läkare göteborg']),
+    ("stockholm läckare göteborg", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),
+    ("stockhlm läckare göteborg", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),
+    ("stockhlm läckare göteborg ", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),  # trailing space
+])
+def test_complete_spelling_correction_multiple_words(session, search_url, query, expected):
+    """
+    Test typeahead with multiple (misspelled) words
+    """
+    typeahead = get_complete(session, search_url, {'q': query})['typeahead']
+    values = []
+    for item in typeahead:
+        values.append(item['value'])
+    # check that expected typeahead is in one of the actual suggestions
+    for e in expected:
+        assert any(v == e for v in values), f"Error: expected: '{e}' not found in '{values}'. Query: {query}"
+
+
+@pytest.mark.parametrize("query, expected", [
+    ("läckare", ['läkare']),
+    ("götteborg", ['göteborg']),
+    ("stokholm", ['stockholm', 'stockholm city']),
+    ("stokholm ", ['stockholm', 'stockholm city']),  # trailing space
+    ("stockhlm", ['stockholm', 'stockholm city']),
+    ("stockhlm ", ['stockholm city']),  # trailing space
+    ("stockholm pythan", ['stockholm python']),
+    ("läkare götteborg", ['läkare göteborg']),
+    ("läkare götteborg", ['läkare göteborg']),
+])
+def test_spelling_correction_with_complete_suggest(session, search_url, query, expected):
+    """
+    Test typeahead with only one (misspelled) word
+    """
+    typeahead = get_complete(session, search_url, {'q': query})['typeahead']
+    values = []
+    for item in typeahead:
+        values.append(item['value'])
+    # check that expected typeahead is in one of the actual suggestions
+    for e in expected:
+        assert any(v == e for v in values), f"Error: expected: '{e}' not found in '{values}'. Query: {query}"
+
+
+@pytest.mark.parametrize("query, expected", [
+    ("götteborg sjukssköterska", ['göteborg sjuksköterska', 'göteborg sjuksköterskan']),
+    ("stokholm lärarre", ['stockholms lärare', 'stockholm lärare', 'stockholm läkare']),
+    ("göteborg sjukssköterska läckare", ['göteborg sjuksköterska lärare', 'göteborg sjuksköterska läkare']),
+    ("göteborg läckare sjukssköterska ", ['göteborg lärare sjuksköterska', 'göteborg läkare sjuksköterska']),
+    ("göteborg läckare sjukssköterska", ['göteborg lärare sjuksköterska', 'göteborg läkare sjuksköterska']),
+    ("läckare götteborg", ['läkare göteborg']),
+    ("läckare göteborg", ['läkare göteborg']),
+    ("stockholm läckare göteborg", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),
+    ("stockhlm läckare göteborg", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),
+    ("stockhlm läckare göteborg ", ['stockholm lärare göteborg', 'stockholm läkare göteborg']),
+])
+def test_spelling_correction_with_phrase_suggest(session, search_url, query, expected):
+    """
+    Test typeahead with several (misspelled) words
+    """
+    typeahead = get_complete(session, search_url, {'q': query})['typeahead']
+    values = []
+    for item in typeahead:
+        values.append(item['value'])
+    # check that expected typeahead is in one of the actual suggestions
+    for e in expected:
+        assert any(v == e for v in values), f"Error: expected: '{e}' not found in '{values}'. Query: {query}"
+
+
+@pytest.mark.parametrize("query, expected", [
+    ("python ", ['python sverige', 'python stockholms län']),
+    ("läkare ", ['läkare sverige', 'läkare svenska'])
+])
+def test_remove_same_word_function(session, search_url, query, expected):
+    """
+    Test when input is word with space, the result will not show the same word
+    """
+    typeahead = get_complete(session, search_url, {'q': query})['typeahead']
+    values = []
+    for item in typeahead:
+        values.append(item['value'])
+    # check that expected typeahead is in one of the actual suggestions
+    for e in expected:
+        assert any(v == e for v in values), f"Error: expected: '{e}' not found in '{values}'. Query: {query}"