+++ tests/api_tests/search/test_search_logic.py
@@ -90,9 +90,9 @@ def test_freetext_two_work_and_two_locations(session, search_url, query, top_id,
 @pytest.mark.skipif(not TEST_USE_STATIC_DATA, reason="depends on a fixed set of ads")
 @pytest.mark.integration
 @pytest.mark.parametrize("query, expected_number, expected_id", [
-    ('Bauhaus Kundtjänst', 1, '23783146'),
+    ('Bauhaus Kundtjänst', 33, '23783146'),
   #  ('Bauhaus Kundtjänst', 33, '23783146'), # when default is changed to 'or'
-    ('Sirius crew', 1, '10537882'),
+    ('Sirius crew', 2, '10537882'),
  #   ('Sirius crew', 2, '10537882'), # when default is changed to 'or'
     ('super', 1, '23801747'),
     ('Säsongande', 1, '23437355'),