+++ tests/integration_tests/search/test_logo_with_workplace_id.py
@@ -9,7 +9,7 @@ from sokannonser.repository import platsannonser
 @pytest.mark.skip(reason="no test data with workplace id")
 @pytest.mark.integration
 def test_fetch_workplace_logo_url_by_ad_id():
-    print('==================', sys._getframe().f_code.co_name, '================== ')
+
 
     app.testing = True
     with app.test_client():
@@ -23,7 +23,7 @@ def test_fetch_workplace_logo_url_by_ad_id():
 @pytest.mark.skip(reason="no test data with workplace id")
 @pytest.mark.integration
 def test_fetch_wp_logo_url_only_org_logo_by_ad_id():
-    print('==================', sys._getframe().f_code.co_name, '================== ')
+
 
     app.testing = True
     with app.test_client():
@@ -37,7 +37,7 @@ def test_fetch_wp_logo_url_only_org_logo_by_ad_id():
 @pytest.mark.skip(reason="no test data with workplace id")
 @pytest.mark.integration
 def test_fetch_missing_logo_url_by_id():
-    print('==================', sys._getframe().f_code.co_name, '================== ')
+
 
     logo_url = platsannonser.get_correct_logo_url('10526669')
     assert logo_url is None